obs-vst: Clean-ups and other small things.
Move files to folders in a More Jim-like structure.
Removed custom code from aeffectx.h and moved it into EditorWidget.h.
Change #includes to reflect the new directory structure.
Reformat files to follow the OBS style guide more closely.
Edited the cmake file to allow for Steinburg headers to be used instead of the bundled one.
