+++ CMakeLists.txt
@@ -4,6 +4,21 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_CURRENT_SOURCE_DIR}")
 
 find_package(Qt5Widgets REQUIRED)
 
+option(VST_USE_BUNDLED_HEADERS "Build with Bundled Headers" ON)
+
+if(VST_USE_BUNDLED_HEADERS)
+	message(STATUS "Using the built-in VST header.")
+	include_directories(vst_header)
+	set(vst_HEADER
+		vst_header/aeffectx.h)
+else()
+	set(VST_INCLUDE_DIR "" CACHE PATH "Path to Steinburg headers (e.g. C:/VST3 SDK/pluginterfaces/vst2.x)")
+	message(WARNING "You should only use the steinburg headers for debugging or local builds. It is illegal to distribute the steinburg headers with anything, and possibly against the GPL to distribute the binaries from the resultant compile.")
+	include_directories(${VST_INCLUDE_DIR})
+	set(vst_HEADER
+		${VST_INCLUDE_DIR}/aeffectx.h)
+endif()
+
 if(APPLE)
     find_library(FOUNDATION_FRAMEWORK Foundation)
     find_library(COCOA_FRAMEWORK Cocoa)
@@ -16,23 +31,23 @@ set(obs-vst_SOURCES
 
 if(APPLE)
     list(APPEND obs-vst_SOURCES
-            VSTPlugin-osx.mm
-            EditorWidget-osx.mm)
+            mac/VSTPlugin-osx.mm
+            mac/EditorWidget-osx.mm)
 elseif(WIN32)
     list(APPEND obs-vst_SOURCES
-            VSTPlugin-win.cpp
-            EditorWidget-win.cpp)
+            win/VSTPlugin-win.cpp
+            win/EditorWidget-win.cpp)
 endif()
 
-set(obs-vst_HEADERS
-        vst-plugin-callbacks.hpp
-        EditorWidget.h
-        VSTPlugin.h
-        aeffectx.h)
+list(APPEND obs-vst_HEADERS
+        headers/vst-plugin-callbacks.hpp
+        headers/EditorWidget.h
+        headers/VSTPlugin.h)
 
 add_library(obs-vst MODULE
         ${obs-vst_SOURCES}
-        ${obs-vst_HEADERS})
+        ${obs-vst_HEADERS}
+		${vst-HEADER})
 
 target_link_libraries(obs-vst
         Qt5::Widgets
