+++ EditorWIdget.cpp
@@ -1,4 +1,4 @@
-#include "EditorWidget.h"
+#include "headers/EditorWidget.h"
 
 EditorWidget::EditorWidget(QWidget *parent, VSTPlugin *plugin) : QWidget(parent), plugin(plugin) {
 	setWindowFlags(this->windowFlags() |= Qt::MSWindowsFixedSizeDialogHint);
